requirements.txt
beautifulsoup4==4.9.3
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
google==3.0.0
google-cloud==0.34.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
soupsieve==2.1
SQLAlchemy==1.3.20
webapp2==2.5.2
Werkzeug==1.0.1
jws~=0.1.3
pip~=20.3.3
wheel~=0.35.1
rsa~=4.6
pyasn1~=0.4.8
six~=1.15.0
gcloud~=0.17.0
httplib2~=0.18.1
oauth2client~=3.0.0
requests~=2.11.1
Pyrebase~=3.0.27
setuptools~=44.1.1
jmespath~=0.10.0
cffi~=1.14.4
Pillow~=8.0.1
numpy~=1.19.4
argh~=0.26.2
pytz~=2020.5
botocore~=1.19.44
boto3~=1.16.44
s3transfer~=0.3.3
redis~=3.5.3
pandas~=1.2.0
urllib3~=1.26.2
python-dateutil~=2.8.1
pilkit~=2.0
cssselect2~=0.4.1
tinycss2~=1.1.0
defusedxml~=0.6.0
cairocffi~=1.2.0
webencodings~=0.5.1
pycparser~=2.20
legacy~=0.1.6
CairoSVG~=2.5.0
libcst~=0.3.16
PyYAML~=5.3.1
blinker~=1.4
cachetools~=4.2.0
certifi~=2020.12.5
matplotlib~=3.3.3
pyarrow~=2.0.0
chardet~=4.0.0
attrs~=20.3.0
idna~=2.10
torch~=1.7.1
scipy~=1.6.0
torchvision~=0.8.2